+++ core/components/romanescobackyard/docs/changelog.md
@@ -3,6 +3,7 @@
 ## Backyard 1.0.0-beta17
 Released on ...
 
+- Modify lexicon strings of Qredits component
 - Include CSS of individual components when assessing critical CSS
 - Use separate setting for critical CSS path
 - Add lexicon keys for overview filters